only do maintenance when triggered by the service

+++ bin/pastmonth.sh
@@ -8,30 +8,35 @@ pushd "${HERE}" >/dev/null || exit 1
 # shellcheck disable=SC1091
 source ./constants.sh
 
-CURRENT_EPOCH=$(date +'%s')
-# do some maintenance
-# shellcheck disable=SC2154
-echo "${db_full_path} re-indexing... "
-sqlite3 "${db_full_path}" "REINDEX;"
 
-if [ $? -eq 0 ]; then
-    echo -n "${db_full_path} integrity check:   "
-    chk_result=$(sqlite3 "${db_full_path}" "PRAGMA integrity_check;")
-    echo " ${chk_result}"
-    if [ "${chk_result}" == "ok" ]; then
-        echo "${db_full_path} copying... "
-        # shellcheck disable=SC2154
-        cp "${db_full_path}" "${database_path}/backup/"
+if [ "${1}" == "-"]; then
+    # do some maintenance
 
-        # Keep upto 10 years of data
-        echo "${db_full_path} vacuuming... "
-        PURGE_EPOCH=$(echo "${CURRENT_EPOCH} - (3660 * 24 * 3600)" | bc)
-        sqlite3 "${db_full_path}" \
-            "DELETE FROM mains WHERE sample_epoch < ${PURGE_EPOCH};"
-        sqlite3 "${db_full_path}" \
-            "DELETE FROM production WHERE sample_epoch < ${PURGE_EPOCH};"
-        sqlite3 "${db_full_path}" \
-            "DELETE FROM charger WHERE sample_epoch < ${PURGE_EPOCH};"
+    CURRENT_EPOCH=$(date +'%s')
+
+    # shellcheck disable=SC2154
+    echo "${db_full_path} re-indexing... "
+    sqlite3 "${db_full_path}" "REINDEX;"
+
+    if [ $? -eq 0 ]; then
+        echo -n "${db_full_path} integrity check:   "
+        chk_result=$(sqlite3 "${db_full_path}" "PRAGMA integrity_check;")
+        echo " ${chk_result}"
+        if [ "${chk_result}" == "ok" ]; then
+            echo "${db_full_path} copying... "
+            # shellcheck disable=SC2154
+            cp "${db_full_path}" "${database_path}/backup/"
+
+            # Keep upto 10 years of data
+            echo "${db_full_path} vacuuming... "
+            PURGE_EPOCH=$(echo "${CURRENT_EPOCH} - (3660 * 24 * 3600)" | bc)
+            sqlite3 "${db_full_path}" \
+                "DELETE FROM mains WHERE sample_epoch < ${PURGE_EPOCH};"
+            sqlite3 "${db_full_path}" \
+                "DELETE FROM production WHERE sample_epoch < ${PURGE_EPOCH};"
+            sqlite3 "${db_full_path}" \
+                "DELETE FROM charger WHERE sample_epoch < ${PURGE_EPOCH};"
+        fi
     fi
 fi
